python-flask/requirements.txt
Flask==2.3.2
web3==6.5.0
eth_account=0.9.0
python-dotenv=1.0.0
